Refactored towards more flexible verification
