+++ Immutable/Apex.Analyzers.Immutable.Test/ApexAnalyzersImmutableUnitTests.cs
@@ -8,7 +8,7 @@ using Xunit;
 
 namespace Apex.Analyzers.Immutable.Test
 {
-    public class UnitTest : CodeFixVerifier
+    public class UnitTest
     {
 
         //No diagnostics expected to show up
@@ -41,10 +41,9 @@ namespace Apex.Analyzers.Immutable.Test
                         }
             };
 
-            VerifyCSharpDiagnostic(test, expected);
-
             var fixtest = test.Replace("private int x", "private readonly int x");
-            VerifyCSharpFix(test, fixtest);
+
+            VerifyCSharpFix(test, new[] { expected }, fixtest);
         }
 
         [Fact]
@@ -84,10 +83,8 @@ namespace Apex.Analyzers.Immutable.Test
                         }
             };
 
-            VerifyCSharpDiagnostic(test, expected);
-
             var fixtest = test.Replace("public int x", "public readonly int x");
-            VerifyCSharpFix(test, fixtest);
+            VerifyCSharpFix(test, new[] { expected }, fixtest);
         }
 
         [Fact]
@@ -150,10 +147,8 @@ namespace Apex.Analyzers.Immutable.Test
                         }
             };
 
-            VerifyCSharpDiagnostic(test, expected);
-
             var fixtest = test.Replace("private int x {get; set;}", "private int x {get; }");
-            VerifyCSharpFix(test, fixtest);
+            VerifyCSharpFix(test, new[] { expected }, fixtest);
         }
 
         [Fact]
@@ -1020,14 +1015,14 @@ namespace Apex.Analyzers.Immutable.Test
             VerifyCSharpDiagnostic(test, expected);
         }
 
-        protected override CodeFixProvider GetCSharpCodeFixProvider()
+        private void VerifyCSharpDiagnostic(string source, params DiagnosticResult[] expected)
         {
-            return new ApexAnalyzersImmutableCodeFixProvider();
+            CSharpCodeFixVerifier<ApexAnalyzersImmutableAnalyzer, ApexAnalyzersImmutableCodeFixProvider>.VerifyAnalyzer(source, expected);
         }
 
-        protected override DiagnosticAnalyzer GetCSharpDiagnosticAnalyzer()
+        private void VerifyCSharpFix(string oldSource, DiagnosticResult[] expected, string newSource)
         {
-            return new ApexAnalyzersImmutableAnalyzer();
+            CSharpCodeFixVerifier<ApexAnalyzersImmutableAnalyzer, ApexAnalyzersImmutableCodeFixProvider>.VerifyCodeFix(oldSource, expected, newSource);
         }
 
         private string GetCode(string code, string namesp = "ConsoleApplication1")
